Reorganize data download strategies
