+++ lib/transports/__es__/_data.js
@@ -41,6 +41,41 @@ class Data {
     return searchAfterResultSet(this, callback)
   }
 
+  /**
+   * Search via size/from, not scroll API.
+   *
+   * This is for OpenSearch Serverless where it does not (yet) support _search?scroll API.
+   * See https://docs.aws.amazon.com/opensearch-service/latest/developerguide/serverless-genref.html#serverless-operations.
+   */
+  async _getDataOffsetLimit (limit, offset, searchBody, callback) {
+    searchBody.size = limit
+    searchBody.from = this.currentOffset ? this.currentOffset : offset
+
+    const additionalParams = this.paramsToString(this.parent.options[`${this.options.type}-params`] || this.parent.options.params, '&')
+
+    const uri = `${this.base.url}/_search${additionalParams}`
+
+    const searchRequest = {
+      uri,
+      method: 'GET',
+      body: jsonParser.stringify(searchBody)
+    }
+    aws4signer(searchRequest, this.parent).then((res) => {
+      this.baseRequest(searchRequest, (err, response) => {
+        err = this.handleError(err, response)
+        if (err) {
+          return callback(err, [])
+        }
+
+        const body = jsonParser.parse(response.body, this.parent)
+        const hits = _.get(body, 'hits.hits', [])
+
+        this.currentOffset = offset + limit
+        return callback(null, hits)
+      })
+    }).catch(callback)
+  }
+
   async _getDataScroll (limit, offset, searchBody, callback) {
     // this allows dumps to be resumed of failed pre-maturely
     // ensure scrollTime is set to a fair amount to prevent
@@ -115,6 +150,10 @@ class Data {
     // Use search_after/pit if specified in options
     if (this.parent.options.searchAfter) {
       return this._getDataPit(searchBody, callback)
+    }
+
+    if (this.ESDistribution === "opensearch-serverless") {
+      return this._getDataOffsetLimit(limit, offset, searchBody, callback)
     }
 
     return _getDataScroll(limit, offset, searchBody, callback)
