+++ lib/transports/__es__/index.js
@@ -8,5 +8,4 @@ module.exports = {
   Template: require('./_template'),
   Script: require('./_script'),
   Data: require('./_data'),
-  Serverless: require('./_serverless')
 }
