+++ lib/transports/elasticsearch.js
@@ -14,11 +14,10 @@ const {
   Policy,
   Setting,
   Template,
-  Serverless,
   Script
 } = require('./__es__')
 
-class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting, Template, Script, Data, Serverless) {
+class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting, Template, Script, Data) {
   constructor (parent, url, options) {
     super()
     this.base = parseBaseURL(url, options)
@@ -133,11 +132,7 @@ class elasticsearch extends Many(Base, Alias, Analyzer, Mapping, Policy, Setting
       if (err) { return callback(err) }
 
       if (type === 'data') {
-        if (this.ESDistribution === "opensearch-serverless") {
-          this.getOpenSearchServerlessData(limit, offset, callback)
-        } else {
-          this.getData(limit, offset, callback)
-        }
+        this.getData(limit, offset, callback)
       } else if (type === 'mapping') {
         this.getMapping(limit, offset, callback)
       } else if (type === 'analyzer' || type === 'settings' || type === 'index') {
